separate Ask the Oracle enum for use by other cmds
